Structured the code
