+++ my-app/src/App.js
@@ -1,48 +1,22 @@
 import React from "react";
-import { Navbar, Container, NavbarBrand } from "react-bootstrap/";
-import "bootstrap/dist/css/bootstrap.min.css";
-import "./App.css";
-import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
 
-import Login from "./components/Login.component";
-import SignUp from "./components/Signup.component";
+import "./App.css";
+import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
+
+import Login from "./components/Login/Login.component";
+import SignUp from "./components/Login/Signup.component";
+import Main from "./components/Main/Main.component";
 
 function App() {
   return (
     <Router>
       <div className="App">
-        <Navbar collapseOnSelect expand="lg" bg="light">
-          <Container>
-            <NavbarBrand>
-              <Link to={"/sign-in"}>RemoteStack</Link>
-            </NavbarBrand>
-            <Navbar.Toggle aria-controls="responsive-navbar-nav" />
-            <Navbar.Collapse id="responsive-navbar-nav">
-              <ul className="navbar-nav ml-auto">
-                <li className="nav-item">
-                  <Link className="nav-link" to={"/sign-in"}>
-                    Sign in
-                  </Link>
-                </li>
-                <li className="nav-item">
-                  <Link className="nav-link" to={"/sign-up"}>
-                    Sign up
-                  </Link>
-                </li>
-              </ul>
-            </Navbar.Collapse>
-          </Container>
-        </Navbar>
-
-        <div className="outer">
-          <div className="inner">
-            <Switch>
-              <Route exact path="/" component={Login} />
-              <Route path="/sign-in" component={Login} />
-              <Route path="/sign-up" component={SignUp} />
-            </Switch>
-          </div>
-        </div>
+        <Switch>
+          <Route exact path="/" component={Login} />
+          <Route exact path="/sign-in" component={Login} />
+          <Route exact path="/sign-up" component={SignUp} />
+          <Route exact path="/dashbord" component={Main} />
+        </Switch>
       </div>
     </Router>
   );
