+++ my-app/src/components/Login/Navbar.css
@@ -1,0 +1,13 @@
+.outer {
+  width: 100%;
+  height: 100%;
+}
+
+.navbar-light {
+  background-color: #ffffff;
+  box-shadow: 0px 14px 80px rgba(34, 35, 58, 0.2);
+}
+
+.custom-control-label {
+  font-weight: 400;
+}
