+++ my-app/src/components/Main/Main.component.js
@@ -1,0 +1,12 @@
+import { render } from "@testing-library/react";
+import React, { Component } from "react";
+import "./Main.css";
+export default class Main extends Component {
+  render() {
+    return (
+      <div className="Main">
+        <h1>Hello</h1>
+      </div>
+    );
+  }
+}
