+++ my-app/src/components/Main/Main.css
@@ -1,0 +1,3 @@
+body {
+  background-color: white;
+}
